test: move pydantic tests from test_complex_types
